fix(ui): extract shared case sort module
